Update README.md
Updated readme a bit.  This will be basis for growing a app and maybe even a JavaFX framework as none are out their that are kept up and maintained.  I think this is critical to the Clojure community to further our work for better dev environments and tools. I hope that we can utilize JavaFX to do better REPL dev and create visualizations as talked about in the REPL talk from https://vimeo.com/223309989